LU-10171 lmv: avoid gratuitous 64-bit modulus
Fix the pct() calculation to use unsigned long arguments, since this
is what callers use.  Remove duplicate pct() definition in lproc_mdc.

Don't do a 64-bit modulus of the LNet NID to find the starting MDT
index when this isn't really needed.

Similarly, don't compute the FLD cache usage percentage for a debug
message that is never used.

Fixes: 9b924e86b27d ("LU-10171 headers: define pct(a,b) once")
Reviewed-on: https://review.whamcloud.com/33922
Tested-by: Jenkins
Tested-by: Maloo <maloo@whamcloud.com>

+++ lustre/lmv/lmv_obd.c
@@ -1373,8 +1373,11 @@ static int lmv_select_statfs_mdt(struct lmv_obd *lmv, __u32 flags)
 			break;
 
 		if (LNET_NETTYP(LNET_NIDNET(lnet_id.nid)) != LOLND) {
+			/* We dont need a full 64-bit modulus, just enough
+			 * to distribute the requests across MDTs evenly.
+			 */
 			lmv->lmv_statfs_start =
-				lnet_id.nid % lmv->desc.ld_tgt_count;
+				(u32)lnet_id.nid % lmv->desc.ld_tgt_count;
 			break;
 		}
 	}
